% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseparameters.R
\name{baseparameters}
\alias{baseparameters}
\title{Get a standard parameter set}
\usage{
baseparameters(
  wInf = 10^seq(log10(4), log10(30000), length.out = 18),
  kappa = 1.3,
  h = 20
)
}
\arguments{
\item{wInf}{Array of asymptotic sizes to use}

\item{kappa}{Carrying capacity of the resource}

\item{h}{Factor for the maximum consumption rate}
}
\value{
A structure with parameters
}
\description{
Get a standard parameter set
}
\details{
See the source code for description of fields
}
