% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runManual.R
\name{plotResults}
\alias{plotResults}
\title{Plot the results from the food web model or the trait-based model.}
\usage{
plotResults(S)
}
\arguments{
\item{S}{Output from a previous simulation to plot}
}
\value{
-
}
\description{
\strong{Left column}.
The left column shows output as a function of body weight of individuals.
Top panel: biomass spectra. Thick solid line is the community spectrum, thin lines
are the individual species or asymptotic size groups. The dashed line is the theoretical
expectation. The thick dashed line is the resource spectru.
Middle panel: feeding level, i.e. consumption rate divided by maximum consumption rate.
The dashed line is the theoretical expectation and the dotted line is the level
where the fish would starve.
Lower panel: mortality. Lines show predation mortality (black) and fishing mortality (red). 
Dashed line is the theoretical expectation. Blue circles is the constant background 
mortality on each species or asymptotic size group.
\strong{Right column}.
The right column shows out as a functio of the asymptotic size of species
(from the food web model) or asymptotic size groups (from the trait-based model).
Top panel: Spawning stock biomass
Middle panel: Life time reproroductive output per egg in the absence of stock-recruitment
relation. This is a measure of the recruitment of the species or asymptotic size group: 
a value below 1 indicates that the stock is collapsed (dotted line).
Lower panel: fisheries yield.
}
\examples{
plotResults( runTraitBasedModel() )

}
