% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runManual.R
\name{runTraitBasedModel}
\alias{runTraitBasedModel}
\title{Run the community spectrum simulator manually}
\usage{
runTraitBasedModel(nSpecies = 18, F0 = 0.3 + 0 * (1:nSpecies), S = NA,
  param = baseparameters(10^seq(log10(4), log10(30000), length.out =
  nSpecies), kappa = 1.3, h = 20))
}
\arguments{
\item{nSpecies}{Number of asymptotic size groups ("species").  Only relevant for trait-based model.}

\item{F0}{Array with fishing mortality on each species}

\item{S}{Output from a previous simulation to use as initial conditions}

\item{param}{Set of parameters to use}
}
\value{
A structure with the output of the simulation with the main fields being:

- t: The time steps where the solution is saved

- nSpecies: no. of asymptotic size groups

- w: the weight classes

- g: growth rate as a function of: time, species, weight

- f: feeding level (consumption divided by max consumption) as a function of: time, species, weight

- M2R: predation mortality on resource

- M2, Ms, Fin: mortality from predation, starvation and fishing

- Ntot: community spectrum as a function of time and weight

- N: species spectra as a function of time, species and weight

- SSB: spawning stock biomass as a function of time and species

- Yield: fisheries yield as a function of time and species
}
\description{
Run the community spectrum simulator manually
}
\examples{
runTraitBasedModel()


}
