% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runManual.R
\docType{data}
\name{shinyui}
\alias{shinyui}
\title{Shiny ui}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.
}
\usage{
shinyui
}
\value{

}
\description{
Shiny ui
}
\examples{

}
\keyword{datasets}
